Backend/requirements.txt
flask
flask-cors
mistralai
groq
openai
python-dotenv
pymupdf